+++ crowddata/menu_map.py
@@ -1,4 +1,4 @@
-from crowddata.mixins.vem.menu.closable_embed import ClosableEmbedMenu
+from crowddata.mixins.adpem.menu.closable_embed import ClosableEmbedMenu
 
 crowddata_menu_map = {
     ClosableEmbedMenu.MENU_TYPE: ClosableEmbedMenu
