+++ crowddata/mixins/adpem/view/show_stats.py
